update order controller

+++ src/main/java/eu/ensup/my_resto/controller/CartController.java
@@ -3,6 +3,7 @@ package eu.ensup.my_resto.controller;
 import eu.ensup.my_resto.model.ItemDTO;
 import eu.ensup.my_resto.model.OrderItemsDTO;
 import eu.ensup.my_resto.service.ItemService;
+import org.aspectj.weaver.ast.Var;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -13,6 +14,7 @@ import javax.servlet.http.HttpSession;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicReference;
+import java.util.stream.Collectors;
 
 @RestController
 public class CartController {
@@ -23,34 +25,6 @@ public class CartController {
     @Autowired
     private ItemService itemService ;
 
-    @PostMapping()
-
-    @PutMapping("/cart")
-    public ResponseEntity<Void> updateCartItem(@RequestBody OrderItemsDTO orderItemDTO, HttpServletRequest req){
-        HttpSession session = req.getSession();
-        Object cart = session.getAttribute("cart");
-        if(cart == null){
-            orderItemsDTOS.clear();
-            cartTotal = 0.0;
-        }
-        AtomicReference<Double> total = new AtomicReference<>(0.0);
-        orderItemsDTOS.forEach(orderItem -> {
-            if(orderItemDTO.getItem().getId().equals(orderItem.getItem().getId())){
-                ItemDTO itemDTO = itemService.get(orderItem.getItem().getId());
-                orderItem.setQuantity(orderItemDTO.getQuantity());
-                orderItem.setItem(itemDTO);
-                total.updateAndGet(value -> value + (itemDTO.getPrice() * orderItem.getQuantity()));
-            }
-        });
-
-        cartTotal = total.get();
-        session.setAttribute("cart", orderItemsDTOS);
-        session.setAttribute("total", cartTotal);
-        System.out.println(session.getAttribute("total"));
-
-        return  ResponseEntity.ok().build();
-    }
-
     @PostMapping("/addCart")
     public ResponseEntity<HttpStatus> addToCart(@RequestBody OrderItemsDTO orderItemDTO, HttpServletRequest req) {
         HttpSession session = req.getSession();
@@ -60,10 +34,25 @@ public class CartController {
             cartTotal = 0.0;
         }
 
-        ItemDTO itemDTO = itemService.get(orderItemDTO.getItem().getId());
-        cartTotal += (itemDTO.getPrice() * orderItemDTO.getQuantity());
-        orderItemDTO.setItem(itemDTO);
-        orderItemsDTOS.add(orderItemDTO);
+        AtomicReference<Boolean> exists = new AtomicReference<>(false);
+        //  check if item exists
+        orderItemsDTOS.forEach(orderItem -> {
+            if (orderItemDTO.getItem().getId().equals(orderItem.getItem().getId())) { // if exist update with total
+                exists.set(true);
+                ItemDTO itemDTO = itemService.get(orderItem.getItem().getId());
+                orderItem.setQuantity(orderItemDTO.getQuantity());
+                orderItem.setItem(itemDTO);
+                cartTotal += (itemDTO.getPrice() * orderItemDTO.getQuantity());
+            }
+        });
+
+        // item does not exist add it
+        if(!exists.get()){
+            ItemDTO itemDTO = itemService.get(orderItemDTO.getItem().getId());
+            cartTotal += (itemDTO.getPrice() * orderItemDTO.getQuantity());
+            orderItemDTO.setItem(itemDTO);
+            orderItemsDTOS.add(orderItemDTO);
+        }
 
         session.setAttribute("cart", orderItemsDTOS);
         session.setAttribute("total", cartTotal);
@@ -78,11 +67,14 @@ public class CartController {
         if(cart == null){
             orderItemsDTOS.clear();
             cartTotal = 0.0;
-        } else{
-            orderItemsDTOS.removeIf(item-> {
-                cartTotal -= (item.getItem().getPrice() * item.getQuantity());
-                return item.getItem().getId().equals(id);
-            });
+        }
+
+        for(int i = 0;  i <orderItemsDTOS.size(); i++){
+            if(orderItemsDTOS.get(i).getItem().getId().equals(id)) {
+                cartTotal -= (orderItemsDTOS.get(i).getItem().getPrice() * orderItemsDTOS.get(i).getQuantity());
+                orderItemsDTOS.remove(i);
+                break;
+            }
         }
 
         session.setAttribute("cart", orderItemsDTOS);
